pyproject.toml
[project]
name = "rag-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.15",
    "google-genai>=1.24.0",
    "python-dotenv>=1.0.0",
    "langchain_text_splitters>=0.1.0",
    "sentence-transformers>=5.0.0",
    "python-docx>=1.1.0",
]
